+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 - remove `stl_debug` (replaced by `stl_hardening`)
 - add `stl_hardening` (default off fast extensive debug debug_with_broken_abi)
-
+- remove `analyzer=taint`
 
 ## 2024-03-23